# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Wildflower Fundraising CRM API (Airtable-aligned schema)
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /people:
    get:
      operationId: listPeople
      tags: [people]
      parameters:
        - { name: search, in: query, schema: { type: string } }
        - { name: deceased, in: query, schema: { type: boolean } }
        - { name: regionId, in: query, schema: { type: string } }
        # Presence filters on computed rollup fields. `has` keeps rows with a
        # value, `blank` keeps rows with none. Omit for no constraint.
        - { name: lifetimeGivingPresence, in: query, required: false, description: "Rollup presence filter on lifetime giving (`has` = >0, `blank` = none).", schema: { type: string, enum: [has, blank] } }
        - { name: lastGiftPresence, in: query, required: false, description: "Rollup presence filter on most-recent gift date (`has` = any gift, `blank` = none).", schema: { type: string, enum: [has, blank] } }
        - { name: openAsksPresence, in: query, required: false, description: "Rollup presence filter on open opportunity count (`has` = >0, `blank` = none).", schema: { type: string, enum: [has, blank] } }
        - { name: activeAffiliationPresence, in: query, required: false, description: "Presence filter on current funder/organization roles (`has` = any current role, `blank` = none).", schema: { type: string, enum: [has, blank] } }
        - name: capacityRating
          in: query
          description: |
            Capacity-rating slugs (see CapacityRating). Accepts the
            literal `__blank__` to match rows with no capacity rating set.
          style: form
          explode: false
          schema: { type: array, items: { type: string } }
        - name: connectionStatus
          in: query
          description: |
            Connection-status slugs (see ConnectionStatus). Accepts the
            literal `__blank__` to match rows with no connection status set.
          style: form
          explode: false
          schema: { type: array, items: { type: string } }
        - name: enthusiasm
          in: query
          style: form
          explode: false
          schema: { type: array, items: { type: string } }
        - name: ownerUserId
          in: query
          description: |
            Owner user-id values. Accepts the literal `__blank__` to match
            rows with no owner assigned.
          style: form
          explode: false
          schema: { type: array, items: { type: string } }
        - name: priority
          in: query
          description: |
            Filter to people whose `priority` tier is in the given set
            (top/high/medium/low). Multi-value: repeat or comma-separate.
            Accepts the literal `__blank__` to match rows with no priority set.
          style: form
          explode: false
          schema: { type: array, items: { type: string } }
        - name: regionIds
          in: query
          description: |
            Filter to people whose `regionIds` array overlaps the given set
            (any selected region appears in the person's regions). Multi-value:
            repeat or comma-separate. Distinct from `regionId` which filters
            on `currentHomeRegionId`.
          style: form
          explode: false
          schema: { type: array, items: { type: string } }
        - { $ref: "#/components/parameters/Limit" }
        - { $ref: "#/components/parameters/Page" }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/PersonList" } } } }
    post:
      operationId: createPerson
      tags: [people]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/CreatePersonBody" } } } }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/Person" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
  /people/{id}:
    get:
      operationId: getPerson
      tags: [people]
      parameters: [ { $ref: "#/components/parameters/IdPath" } ]
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/PersonDetail" } } } }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      operationId: updatePerson
      tags: [people]
      parameters: [ { $ref: "#/components/parameters/IdPath" } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/UpdatePersonBody" } } } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Person" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      operationId: deletePerson
      tags: [people]
      parameters: [ { $ref: "#/components/parameters/IdPath" } ]
      responses: { "204": { description: Deleted } }

  # ─── People-entity-roles ───────────────────────────────────────────────────
  /people-entity-roles:
    get:
      operationId: listPeopleEntityRoles
      tags: [people-entity-roles]
      parameters:
        - { name: personId, in: query, schema: { type: string } }
        - { name: organizationId, in: query, schema: { type: string } }
        - { name: paymentIntermediaryId, in: query, schema: { type: string } }
        - { name: householdId, in: query, schema: { type: string } }
        - { $ref: "#/components/parameters/Limit" }
        - { $ref: "#/components/parameters/Page" }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/PeopleEntityRoleList" } } } }
    post:
      operationId: createPeopleEntityRole
      tags: [people-entity-roles]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/CreatePeopleEntityRoleBody" } } } }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/PeopleEntityRole" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
  /people-entity-roles/{id}:
    patch:
      operationId: updatePeopleEntityRole
      tags: [people-entity-roles]
      parameters: [ { $ref: "#/components/parameters/IdPath" } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/UpdatePeopleEntityRoleBody" } } } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/PeopleEntityRole" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      operationId: deletePeopleEntityRole
      tags: [people-entity-roles]
      parameters: [ { $ref: "#/components/parameters/IdPath" } ]
      responses: { "204": { description: Deleted } }

  # ─── Emails ────────────────────────────────────────────────────────────────
...
  /people/bulk-update:
    post:
      operationId: bulkUpdatePeople
      tags: [people]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/BulkUpdatePeopleBody" } } } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/BulkUpdateResult" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
...
  /people/merge:
    post:
      operationId: mergePeople
      tags: [people]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/MergePeopleBody" } } } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/MergeResult" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
...
components:
  parameters:
    Limit: { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 10000, default: 50 } }
    Page:  { name: page,  in: query, schema: { type: integer, minimum: 1, default: 1 } }
    IdPath: { name: id, in: path, required: true, schema: { type: string } }
  responses:
    NotFound:
      description: Not found
      content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
    BadRequest:
      description: Bad request (validation error)
      content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
...
  schemas:
...
    Person:
      type: object
      required: [id, deceased, newsletter, unsubscribedToNewsletter, anonymous, createdAt, updatedAt]
      properties:
        id:                       { type: string }
        prefix:                   { type: string, nullable: true }
        firstName:                { type: string, nullable: true }
        nickname:                 { type: string, nullable: true }
        middleName:               { type: string, nullable: true }
        lastName:                 { type: string, nullable: true }
        suffix:                   { type: string, nullable: true }
        fullName:                 { type: string, nullable: true }
        pronouns:                 { allOf: [ { $ref: "#/components/schemas/Pronouns" } ], nullable: true }
        deceased:                 { type: boolean }
        currentHomeRegionId:      { type: string, nullable: true }
        details:                  { type: string, nullable: true }
        ownerUserId:                    { type: string, nullable: true }
        tags:                     { type: string, nullable: true }
        lastContacted:            { type: string, format: date, nullable: true }
        interactionCount:         { type: integer, nullable: true }
        linkedin:                 { type: string, nullable: true }
        x:                        { type: string, nullable: true }
        facebook:                 { type: string, nullable: true }
        instagram:                { type: string, nullable: true }
        aboutMe:                  { type: string, nullable: true }
        youtube:                  { type: string, nullable: true }
        website:                  { type: string, nullable: true }
        interestsThematic:        { type: array, items: { type: string }, nullable: true }
        interestsAges:            { type: array, items: { type: string }, nullable: true }
        interestsGovModels:       { type: array, items: { type: string }, nullable: true }
        regionIds:                { type: array, items: { type: string }, nullable: true }
        newsletter:               { type: boolean }
        unsubscribedToNewsletter: { type: boolean }
        childrenAtWf:             { type: string, nullable: true }
        meetingLink:              { type: string, nullable: true }
        assistantPersonId:        { type: string, nullable: true }
        capacityRating:           { allOf: [ { $ref: "#/components/schemas/CapacityRating" } ], nullable: true }
        netWorth:                 { type: string, nullable: true, description: "Estimated net worth for this individual. Decimal as string." }
        connectionStatus:         { allOf: [ { $ref: "#/components/schemas/ConnectionStatus" } ], nullable: true }
        enthusiasm:               { allOf: [ { $ref: "#/components/schemas/Enthusiasm" } ], nullable: true }
        priority:                 { allOf: [ { $ref: "#/components/schemas/Priority" } ], nullable: true, description: "Solicitation priority tier (top/high/medium/low). The 'top' band is surfaced as a star on the individuals table and on opportunities/gifts where this person is the individual giver." }
        anonymous:                { type: boolean, description: "When true, hide the person's real name in the UI (shown as 'Anonymous') from everyone except the record owner and admins. UI-only; the name is still stored and returned." }
        # Server-computed aggregates for the individuals list view. NOT
        # writable; computed per-request by the list-people query. Per
        # replit.md: lifetime giving is direct individual gifts + all
        # gifts to households the person belongs to (no splitting).
        lifetimeGiving:           { type: string, nullable: true, readOnly: true, description: "Sum of direct individual gifts + all gifts to households the person belongs to. Decimal as string." }
        mostRecentGiftDate:       { type: string, format: date, nullable: true, readOnly: true, description: "Most recent date_received across direct individual gifts and gifts to households the person belongs to." }
        openOpportunityCount:     { type: integer, nullable: true, readOnly: true, description: "Count of opportunities_and_pledges where this person is the individual giver and status='open'." }
        activeOrganizationNames:  { type: array, items: { type: string }, nullable: true, readOnly: true, description: "Names of organizations the person currently holds a role at (people_entity_roles.current='current')." }
        pastOrganizationNames:    { type: array, items: { type: string }, nullable: true, readOnly: true, description: "Names of organizations the person previously held a role at (people_entity_roles.current='past')." }
        createdAt:                { type: string, format: date-time }
        updatedAt:                { type: string, format: date-time }
    PersonDetail:
      allOf:
        - { $ref: "#/components/schemas/Person" }
        - type: object
          properties:
            roles:        { type: array, items: { $ref: "#/components/schemas/PeopleEntityRole" } }
            emails:       { type: array, items: { $ref: "#/components/schemas/Email" } }
            phoneNumbers: { type: array, items: { $ref: "#/components/schemas/PhoneNumber" } }
            addresses:    { type: array, items: { $ref: "#/components/schemas/Address" } }
    PersonList:
      type: object
      required: [data, pagination]
      properties:
        data:       { type: array, items: { $ref: "#/components/schemas/Person" } }
        pagination: { $ref: "#/components/schemas/Pagination" }
    CreatePersonBody:
      type: object
      properties:
        prefix:                   { type: string }
        firstName:                { type: string }
        nickname:                 { type: string }
        middleName:               { type: string }
        lastName:                 { type: string }
        suffix:                   { type: string }
        fullName:                 { type: string }
        pronouns:                 { $ref: "#/components/schemas/Pronouns" }
        deceased:                 { type: boolean }
        currentHomeRegionId:      { type: string }
        details:                  { type: string }
        ownerUserId:                    { type: string }
        tags:                     { type: string }
        lastContacted:            { type: string, format: date }
        linkedin:                 { type: string }
        x:                        { type: string }
        facebook:                 { type: string }
        instagram:                { type: string }
        aboutMe:                  { type: string }
        youtube:                  { type: string }
        website:                  { type: string }
        interestsThematic:        { type: array, items: { type: string } }
        interestsAges:            { type: array, items: { type: string } }
        interestsGovModels:       { type: array, items: { type: string } }
        regionIds:                { type: array, items: { type: string } }
        newsletter:               { type: boolean }
        unsubscribedToNewsletter: { type: boolean }
        childrenAtWf:             { type: string }
        meetingLink:              { type: string }
        assistantPersonId:        { type: string }
        capacityRating:           { $ref: "#/components/schemas/CapacityRating" }
        netWorth:                 { type: string, description: "Estimated net worth for this individual. Decimal as string." }
        connectionStatus:         { $ref: "#/components/schemas/ConnectionStatus" }
        enthusiasm:               { $ref: "#/components/schemas/Enthusiasm" }
        priority:                 { $ref: "#/components/schemas/Priority" }
        anonymous:                { type: boolean }
    UpdatePersonBody:
      type: object
      properties:
        prefix:                   { type: string, nullable: true }
        firstName:                { type: string, nullable: true }
        nickname:                 { type: string, nullable: true }
        middleName:               { type: string, nullable: true }
        lastName:                 { type: string, nullable: true }
        suffix:                   { type: string, nullable: true }
        fullName:                 { type: string, nullable: true }
        pronouns:                 { allOf: [ { $ref: "#/components/schemas/Pronouns" } ], nullable: true }
        deceased:                 { type: boolean }
        currentHomeRegionId:      { type: string, nullable: true }
        details:                  { type: string, nullable: true }
        ownerUserId:                    { type: string, nullable: true }
        tags:                     { type: string, nullable: true }
        lastContacted:            { type: string, format: date, nullable: true }
        linkedin:                 { type: string, nullable: true }
        x:                        { type: string, nullable: true }
        facebook:                 { type: string, nullable: true }
        instagram:                { type: string, nullable: true }
        aboutMe:                  { type: string, nullable: true }
        youtube:                  { type: string, nullable: true }
        website:                  { type: string, nullable: true }
        interestsThematic:        { type: array, items: { type: string }, nullable: true }
        interestsAges:            { type: array, items: { type: string }, nullable: true }
        interestsGovModels:       { type: array, items: { type: string }, nullable: true }
        regionIds:                { type: array, items: { type: string }, nullable: true }
        newsletter:               { type: boolean }
        unsubscribedToNewsletter: { type: boolean }
        childrenAtWf:             { type: string, nullable: true }
        meetingLink:              { type: string, nullable: true }
        assistantPersonId:        { type: string, nullable: true }
        capacityRating:           { allOf: [ { $ref: "#/components/schemas/CapacityRating" } ], nullable: true }
        netWorth:                 { type: string, nullable: true, description: "Estimated net worth for this individual. Decimal as string." }
        connectionStatus:         { allOf: [ { $ref: "#/components/schemas/ConnectionStatus" } ], nullable: true }
        enthusiasm:               { allOf: [ { $ref: "#/components/schemas/Enthusiasm" } ], nullable: true }
        priority:                 { allOf: [ { $ref: "#/components/schemas/Priority" } ], nullable: true }
        anonymous:                { type: boolean }

    # ─── People-entity-roles ───────────────────────────────────────────────────
...
    PeopleEntityRoleList:
      type: object
      required: [data, pagination]
      properties:
        data:       { type: array, items: { $ref: "#/components/schemas/PeopleEntityRole" } }
        pagination: { $ref: "#/components/schemas/Pagination" }
    CreatePeopleEntityRoleBody:
      type: object
      required: [personId, entityType]
      properties:
        personId:              { type: string }
        entityType:            { $ref: "#/components/schemas/EntityRoleType" }
        organizationId:        { type: string }
        paymentIntermediaryId: { type: string }
        householdId:           { type: string }
        connection:            { $ref: "#/components/schemas/PeopleEntityRoleConnection" }
        notes:                 { type: string }
        externalTitleOrRole:   { type: string }
        current:               { $ref: "#/components/schemas/PeopleRoleCurrent" }
        primaryContact:        { type: boolean }
    UpdatePeopleEntityRoleBody:
      type: object
      properties:
        organizationId:        { type: string, nullable: true }
        paymentIntermediaryId: { type: string, nullable: true }
        householdId:           { type: string, nullable: true }
        connection:            { allOf: [ { $ref: "#/components/schemas/PeopleEntityRoleConnection" } ], nullable: true }
        notes:                 { type: string, nullable: true }
        externalTitleOrRole:   { type: string, nullable: true }
        current:               { $ref: "#/components/schemas/PeopleRoleCurrent" }
        primaryContact:        { type: boolean }

    # ─── Emails / phones / addresses ───────────────────────────────────────────
...
    BulkUpdatePeopleBody:
      type: object
      required: [ids, patch]
      properties:
        ids:   { type: array, items: { type: string }, minItems: 1, maxItems: 1000 }
        patch: { $ref: "#/components/schemas/BulkUpdatePeoplePatch" }
...